+++ 1.12/src/main/java/panszelescik/moreplates/plugins/PluginCalculator.java
@@ -12,8 +12,6 @@ public class PluginCalculator {
 	public static Item amethyst_plate;
 	public static Item enriched_gold_gear;
 	public static Item enriched_gold_plate;
-	public static Item redstone_gear;
-	public static Item redstone_plate;
 	public static Item reinforced_iron_gear;
 	public static Item reinforced_iron_plate;
 	public static Item tanzanite_gear;
@@ -24,8 +22,6 @@ public class PluginCalculator {
 		amethyst_plate = new ItemPlate("amethyst");
 		enriched_gold_gear = new ItemGear("enriched_gold");
 		enriched_gold_plate = new ItemPlate("enriched_gold");
-		redstone_gear =	 new ItemGear("redstone1");
-		redstone_plate = new ItemPlate("redstone1");
 		reinforced_iron_gear = new ItemGear("reinforced_iron");
 		reinforced_iron_plate = new ItemPlate("reinforced_iron");
 		tanzanite_gear = new ItemGear("tanzanite");
@@ -33,7 +29,6 @@ public class PluginCalculator {
 		
 		reg(AMETHYST, amethyst_gear, amethyst_plate);
 		reg(ENRICHED_GOLD, enriched_gold_gear, enriched_gold_plate);
-		reg(REDSTONE_CALCULATOR, redstone_gear, redstone_plate);
 		reg(REINFORCED_IRON, reinforced_iron_gear, reinforced_iron_plate);
 		reg(TANZANITE, tanzanite_gear, tanzanite_plate);
 	}